[REFAC: #60] - Melhora a distribuição das fixtures.
close #60
